requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymupdf>=1.25.5
pdfplumber==0.10.2
pandas==2.1.1
langchain==0.0.312
langchain-openai==0.0.2
tiktoken>=0.5.2,<0.6.0
numpy==1.26.0
supabase==1.2.0
pgvector==0.2.3
python-dotenv==1.0.0
httpx==0.24.1
pillow==10.1.0
matplotlib==3.8.0
opencv-python==4.8.1.78
pydantic-settings>=2.0.0